feat: VM-UNet CVC training + whole-network causal (2x2 grid complete)
- Fix vmamba.py: vectorized cumprod associative scan replaces O(L) serial loop
- Fix vmunet.py: weights_only=False for PyTorch 2.6 checkpoint loading
- Fix train_vmunet_cvc.py: float() conversion in validation loop for AMP BCELoss
- Add experiment14_vmunet_cvc_causal.py: whole-network low-pass on VM-UNet CVC
- Results: Baseline Dice 0.8957, LP Dice 0.2919 (-67.4%)
- 2x2 grid now complete: VM-UNet/ResNet50 x ISIC/CVC

+++ interventions/train_vmunet_cvc.py
@@ -173,7 +173,7 @@ def main():
                 imgs, masks = imgs.to(device), masks.to(device)
                 with autocast:
                     probs = model(imgs)  # already sigmoid-ed
-                val_loss += combined_loss(probs, masks).item()
+                val_loss += combined_loss(probs.float(), masks.float()).item()
                 pred = (probs > 0.5).float()
                 inter = (pred * masks).sum().item()
                 union = pred.sum().item() + masks.sum().item()
